+++ src/habluetooth/wrappers.py
@@ -74,6 +74,16 @@ class HaBleakScannerWrapper(BaseBleakScanner):
         )
 
     @classmethod
+    async def find_device_by_address(
+        cls, device_identifier: str, timeout: float = 10.0, **kwargs: Any
+    ) -> BLEDevice | None:
+        """Find a device by address."""
+        manager = get_manager()
+        return manager.async_ble_device_from_address(
+            device_identifier, True
+        ) or manager.async_ble_device_from_address(device_identifier, False)
+
+    @classmethod
     async def discover(cls, timeout: float = 5.0, **kwargs: Any) -> list[BLEDevice]:
         """Discover devices."""
         return list(get_manager().async_discovered_devices(True))