+++ tests/__init__.py
@@ -1,8 +1,20 @@
-from typing import Any
+import asyncio
+import time
+from contextlib import contextmanager
+from datetime import datetime, timezone
+from functools import partial
+from typing import Any, Generator
+from unittest.mock import patch
 
-from bleak.backends.scanner import AdvertisementData
+from bleak.backends.scanner import AdvertisementData, BLEDevice
+
+utcnow = partial(datetime.now, timezone.utc)
+
+
+_MONOTONIC_RESOLUTION = time.get_clock_info("monotonic").resolution
 
 ADVERTISEMENT_DATA_DEFAULTS = {
+    "local_name": "Unknown",
     "manufacturer_data": {},
     "service_data": {},
     "service_uuids": [],
@@ -12,9 +24,66 @@ ADVERTISEMENT_DATA_DEFAULTS = {
 }
 
 
+BLE_DEVICE_DEFAULTS = {
+    "name": None,
+    "rssi": -127,
+    "details": None,
+}
+
+
 def generate_advertisement_data(**kwargs: Any) -> AdvertisementData:
     """Generate advertisement data with defaults."""
     new = kwargs.copy()
     for key, value in ADVERTISEMENT_DATA_DEFAULTS.items():
         new.setdefault(key, value)
     return AdvertisementData(**new)
+
+
+def generate_ble_device(
+    address: str | None = None,
+    name: str | None = None,
+    details: Any | None = None,
+    rssi: int | None = None,
+    **kwargs: Any,
+) -> BLEDevice:
+    """Generate a BLEDevice with defaults."""
+    new = kwargs.copy()
+    if address is not None:
+        new["address"] = address
+    if name is not None:
+        new["name"] = name
+    if details is not None:
+        new["details"] = details
+    if rssi is not None:
+        new["rssi"] = rssi
+    for key, value in BLE_DEVICE_DEFAULTS.items():
+        new.setdefault(key, value)
+    return BLEDevice(**new)
+
+
+@contextmanager
+def patch_bluetooth_time(mock_time: float) -> Generator[Any, None, None]:
+    """Patch the bluetooth time."""
+    with (
+        patch("habluetooth.base_scanner.monotonic_time_coarse", return_value=mock_time),
+        patch("habluetooth.manager.monotonic_time_coarse", return_value=mock_time),
+        patch("habluetooth.scanner.monotonic_time_coarse", return_value=mock_time),
+    ):
+        yield
+
+
+def async_fire_time_changed(utc_datetime: datetime) -> None:
+    timestamp = utc_datetime.timestamp()
+    loop = asyncio.get_running_loop()
+    for task in list(loop._scheduled):  # type: ignore[attr-defined]
+        if not isinstance(task, asyncio.TimerHandle):
+            continue
+        if task.cancelled():
+            continue
+
+        mock_seconds_into_future = timestamp - time.time()
+        future_seconds = task.when() - (loop.time() + _MONOTONIC_RESOLUTION)
+
+        if mock_seconds_into_future >= future_seconds:
+            task._run()
+            task.cancel()