+++ tests/test_models.py
@@ -4,10 +4,7 @@ import time
 
 from bleak.backends.device import BLEDevice
 
-from habluetooth import (
-    BluetoothServiceInfo,
-    BluetoothServiceInfoBleak,
-)
+from habluetooth import BluetoothServiceInfo, BluetoothServiceInfoBleak
 
 from . import generate_advertisement_data
 
@@ -89,6 +86,7 @@ def test_model_from_scanner():
         "service_uuids": ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"],
         "source": "local",
         "time": now,
+        "tx_power": -127,
     }
 
 
@@ -110,6 +108,7 @@ def test_construct_service_info_bleak():
         advertisement=switchbot_adv,
         connectable=False,
         time=now,
+        tx_power=1,
     )
 
     assert service_info.service_uuids == ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"]
@@ -133,6 +132,7 @@ def test_construct_service_info_bleak():
         "service_uuids": ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"],
         "source": "local",
         "time": now,
+        "tx_power": 1,
     }
 
 
@@ -170,7 +170,14 @@ def test_from_device_and_advertisement_data():
         "service_uuids": ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"],
         "source": "local",
         "time": now_monotonic,
+        "tx_power": -127,
     }
+    assert str(service_info) == (
+        "<BluetoothServiceInfoBleak name=wohand address=44:44:33:11:23:45 rssi=-127 "
+        "manufacturer_data={} service_data={} "
+        "service_uuids=['cba20d00-224d-11e6-9fb8-0002a5d5c51b'] source=local "
+        f"connectable=True time={now_monotonic} tx_power=-127>"
+    )
 
 
 def test_pyobjc_compat():
@@ -205,6 +212,7 @@ def test_pyobjc_compat():
         advertisement=switchbot_adv,
         connectable=False,
         time=now,
+        tx_power=1,
     )
 
     assert service_info.service_uuids == ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"]
@@ -228,4 +236,5 @@ def test_pyobjc_compat():
         "service_uuids": ["cba20d00-224d-11e6-9fb8-0002a5d5c51b"],
         "source": "local",
         "time": now,
+        "tx_power": 1,
     }